@@ -15,32 +15,32 @@
   },
   "private": true,
   "dependencies": {
-    "@angular-devkit/build-angular": "~13.2.2",
-    "@angular/animations": "~13.2.1",
-    "@angular/cdk": "~13.2.1",
-    "@angular/cli": "~13.2.2",
-    "@angular/common": "~13.2.1",
-    "@angular/compiler": "~13.2.1",
-    "@angular/compiler-cli": "~13.2.1",
-    "@angular/core": "~13.2.1",
-    "@angular/forms": "~13.2.1",
-    "@angular/localize": "^13.2.1",
-    "@angular/material": "^13.2.1",
-    "@angular/platform-browser": "~13.2.1",
-    "@angular/platform-browser-dynamic": "~13.2.1",
-    "@angular/router": "~13.2.1",
+    "@angular-devkit/build-angular": "^19.2.8",
+    "@angular/animations": "^19.2.7",
+    "@angular/cdk": "^19.2.10",
+    "@angular/cli": "^19.2.8",
+    "@angular/common": "^19.2.7",
+    "@angular/compiler": "^19.2.7",
+    "@angular/compiler-cli": "^19.2.7",
+    "@angular/core": "^19.2.7",
+    "@angular/forms": "^19.2.7",
+    "@angular/localize": "^19.2.7",
+    "@angular/material": "^19.2.10",
+    "@angular/platform-browser": "^19.2.7",
+    "@angular/platform-browser-dynamic": "^19.2.7",
+    "@angular/router": "^19.2.7",
     "@fortawesome/fontawesome-free": "^5.12.1",
-    "@ngx-translate/core": "^12.1.2",
+    "@ngx-translate/core": "^16.0.4",
     "@ngx-translate/http-loader": "^4.0.0",
     "d3": "^7.0.1",
     "flag-icon-css": "^4.1.7",
     "rxjs": "^6.6.7",
     "tslib": "^2.0.0",
-    "typescript": "~4.5.5",
-    "zone.js": "~0.11.4"
+    "typescript": "~5.8.3",
+    "zone.js": "~0.15.0"
   },
   "devDependencies": {
-    "@angular/language-service": "~13.2.1",
+    "@angular/language-service": "^19.2.7",
     "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "~8.9.4",
